+++ app/src/main/java/com/codingblocks/cbonlineapp/HomeActivity.kt
@@ -54,7 +54,7 @@ class HomeActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelecte
     private fun fetchUser() {
         if (!prefs.SP_ACCESS_TOKEN_KEY.equals("access_token")) {
             info { prefs.SP_ACCESS_TOKEN_KEY }
-            Clients.api.getMe("JWT " + prefs.SP_JWT_TOKEN_KEY).enqueue(retrofitCallback { t, resp ->
+            Clients.api.getMe().enqueue(retrofitCallback { t, resp ->
                 resp?.body()?.let { it ->
                     if (resp.isSuccessful) {
                         val jSONObject = resp.body()!!.getAsJsonObject("data").getAsJsonObject("attributes")
